Refactor: move external links stuff to new files, separate data file, add shorten_type_prefixes to config
Users can now add a str list called sphinx_csharp_shorten_type_prefixes to the config to shorten types
